Fix: Cast DB_PORT to integer and improve password placeholder

+++ Modele/DAO/config.php
@@ -4,7 +4,7 @@
 // DB_TYPE, DB_HOST, DB_PORT, DB_NAME, DB_USER, DB_PASS, DB_CHARSET
 define('DB_TYPE', getenv('DB_TYPE') ?: 'mysql');                 // type de la bd
 define('DB_HOST', getenv('DB_HOST') ?: '127.0.0.1');             // nom du serveur
-define('DB_PORT', getenv('DB_PORT') ?: 3306);                    // port MySQL standard
+define('DB_PORT', (int)(getenv('DB_PORT') ?: 3306));             // port MySQL standard
 define('DB_NAME', getenv('DB_NAME') ?: 'gestion_joueurs');       // nom de la BD par défaut
 define('DB_USER', getenv('DB_USER') ?: 'root');                  // utilisateur par défaut
 define('DB_PASS', getenv('DB_PASS') ?: '');                      // mot de passe par défaut vide
